feat(global): 完成用户提交申请，管理员审核等

+++ src/main/java/com/example/lostfound/entity/LossCommentVO.java
@@ -26,7 +26,7 @@ public class LossCommentVO {
 
     private Integer lostThingId;
 
-    private Long likes;
+    private Integer likes;
 
     private String nickName;
